up to game1 display
