--- js/getElement.js
@@ -1,5 +1,0 @@
-export default (markup) => {
-  let el = document.createElement('div');
-  el.innerHTML = markup;
-  return el.firstChild;
-};
