+++ js/utilities/getElement.js
@@ -1,0 +1,5 @@
+export default (markup) => {
+  let el = document.createElement(`div`);
+  el.innerHTML = markup;
+  return el;
+};
