+++ js/utilities/setScreen.js
@@ -1,4 +1,5 @@
 export default (template) => {
   const mainContainer = document.querySelector(`main.central`);
+  mainContainer.innerHTML = ``;
   mainContainer.appendChild(template);
 };
